requirements.txt
# Comprehensive Requirements File for Hack Digital Transformation Project

# Core ML Dependencies
torch==2.8.0
torchvision==0.23.0
faiss-cpu==1.12.0
numpy>=1.25.0,<2.0.0
pandas==2.3.2
scikit-learn==1.5.1
opencv-python-headless==4.12.0.88
easyocr==1.7.2
Pillow==11.3.0
tqdm==4.67.1
boto3==1.40.35
botocore==1.40.35
python-dotenv==1.1.1
pyyaml==6.0.2
scipy==1.11.0

# FastAPI Core Dependencies
fastapi==0.117.1
uvicorn[standard]==0.36.0
pydantic==2.11.9
pydantic-settings==2.10.1

# Database Dependencies
psycopg2-binary==2.9.10
sqlalchemy==2.0.43

# Redis Dependencies
redis==6.4.0

# Celery Dependencies
celery==5.5.3

# Web and API Dependencies
requests==2.32.5
python-multipart==0.0.20

# Data Processing and Visualization
openpyxl==3.1.5
matplotlib==3.9.2
seaborn==0.13.2
joblib==1.4.2

# Optimization and Experimentation
optuna==4.5.0

# Utility Dependencies
python-slugify==8.0.4
python-magic==0.4.27

# Development and Testing
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0

# Code Quality
ruff==0.13.2